Move Repository to contracts folder
